requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
Pillow>=9.0.0
